+++ src/lib.rs
@@ -2,4 +2,5 @@
 #![doc = include_str!("../README.md")]
 
 pub mod carddav;
+pub mod http;
 pub mod tcp;
